move stories outside of components and reorganize categories
