+++ next.config.js
@@ -16,7 +16,7 @@ const config = {
   poweredByHeader: false,
   eslint: {
     // add background and serverless to the default list of pages for eslint
-    dirs: ['pages', 'components', 'lib', 'background', 'serverless'],
+    dirs: ['pages', 'components', 'lib', 'background', 'serverless', 'stories'],
     ignoreDuringBuilds: skipCodeChecks
   },
   // types are tested separately from the build
