+++ stories/Databases/CardDetails.stories.tsx
@@ -4,13 +4,11 @@ import { rest } from 'msw';
 import type { ReactNode } from 'react';
 import { useRef } from 'react';
 import { Provider } from 'react-redux';
-import type { MockStoreEnhanced } from 'redux-mock-store';
 import { v4 as uuid } from 'uuid';
 
-import Kanban from 'components/common/BoardEditor/focalboard/src/components/kanban/kanban';
-import Table from 'components/common/BoardEditor/focalboard/src/components/table/table';
-import type { RootState } from 'components/common/BoardEditor/focalboard/src/store';
+import CardDetailProperties from 'components/common/BoardEditor/focalboard/src/components/cardDetail/cardDetailProperties';
 import { mockStateStore } from 'components/common/BoardEditor/focalboard/src/testUtils';
+import { CardPropertiesWrapper } from 'components/common/CharmEditor/CardPropertiesWrapper';
 import type { ICurrentSpaceContext } from 'hooks/useCurrentSpace';
 import { CurrentSpaceContext } from 'hooks/useCurrentSpace';
 import { MembersProvider } from 'hooks/useMembers';
@@ -22,15 +20,14 @@ import { createMockBoard, createMockCard } from 'testing/mocks/block';
 import { createMockPage } from 'testing/mocks/page';
 import { generateSchemasForAllSupportedFieldTypes } from 'testing/publicApi/schemas';
 
-import { spaces } from '../../../../.storybook/lib/mockData';
+import { spaces } from '../../.storybook/lib/mockData';
 
 export default {
   title: 'Databases/Composites',
-  component: Kanban
+  component: CardDetailProperties
 };
 
 const firstUserId = uuid();
-const secondUserId = uuid();
 
 const space = spaces[0];
 
@@ -65,12 +62,14 @@ const boardPage: PageMeta = {
   syncWithPageId: null
 };
 
+const view = createTableView({ board });
+
 const card1 = createMockCard(board);
 
 card1.fields.properties = {
   [schema.text.id]: 'First',
   [schema.checkbox.id]: 'true',
-  [schema.date.id]: '{"from":1694001600000}',
+  [schema.date.id]: '{"from":"1695067400713"}',
   [schema.email.id]: 'test1@example.com',
   [schema.multiSelect.id]: [schema.multiSelect.options[0].id, schema.multiSelect.options[1].id],
   [schema.number.id]: 7223,
@@ -79,40 +78,11 @@ card1.fields.properties = {
   [schema.select.id]: schema.select.options[0].id,
   [schema.url.id]: 'https://www.google.com'
 };
-
-const card2 = createMockCard(board);
-
-card2.fields.properties = {
-  [schema.text.id]: 'Second',
-  [schema.checkbox.id]: 'false',
-  [schema.date.id]: '{"from":1694501600000}',
-  [schema.email.id]: 'test2@example.com',
-  [schema.multiSelect.id]: [schema.multiSelect.options[1].id, schema.multiSelect.options[2].id],
-  [schema.number.id]: 8345,
-  [schema.person.id]: secondUserId,
-  [schema.phone.id]: '+1 (345) 8345 345',
-  [schema.select.id]: schema.select.options[1].id,
-  [schema.url.id]: 'https://www.example.com'
-};
-
 const page1 = createMockPage({
   id: card1.id,
   type: 'card',
   title: 'Card 1'
 });
-
-const page2 = createMockPage({
-  id: card2.id,
-  type: 'card',
-  title: 'Card 2'
-});
-const view = createTableView({ board });
-
-view.fields.groupById = schema.select.id;
-view.fields.cardOrder = [card1.id, card2.id];
-view.fields.viewType = 'board';
-view.fields.visibleOptionIds = schema.select.options.map((opt) => opt.id);
-view.fields.visiblePropertyIds = [schema.select.id];
 
 const reduxStore = mockStateStore([], {
   boards: {
@@ -130,12 +100,11 @@ const reduxStore = mockStateStore([], {
   cards: {
     current: '',
     cards: {
-      [card1.id]: card1,
-      [card2.id]: card2
+      [card1.id]: card1
     },
     templates: {}
   }
-}) as MockStoreEnhanced<Pick<RootState, 'boards' | 'views' | 'cards'>>;
+});
 
 function Context({ children }: { children: ReactNode }) {
   // mock the current space since it usually relies on the URL
@@ -157,51 +126,38 @@ function Context({ children }: { children: ReactNode }) {
   );
 }
 
-function voidFunction() {
-  return Promise.resolve();
-}
+export function CardPropsView() {
+  // const dispatch = useAppDispatch();
 
-export function DatabaseKanbanView() {
+  // useEffect(() => {
+  //   dispatch(initialDatabaseLoad({ pageId: board.id }));
+  // }, []);
+
   return (
     <Context>
       <Paper>
-        <div className='focalboard-body'>
-          <Kanban
+        <CardPropertiesWrapper>
+          <CardDetailProperties
             board={board}
-            showCard={voidFunction}
             readOnly={false}
-            cards={[card1, card2]}
-            hiddenGroups={[]}
+            views={[view]}
             activeView={view}
-            groupByProperty={schema.select}
-            addCard={voidFunction}
-            onCardClicked={voidFunction}
-            selectedCardIds={[]}
-            visibleGroups={[
-              { cardPages: [{ card: card1, page: page1 }], cards: [card1], option: schema.select.options[0] },
-              { cardPages: [{ card: card2, page: page2 }], cards: [card2], option: schema.select.options[1] }
-            ]}
+            card={card1}
+            cards={[card1]}
+            pageUpdatedAt={new Date().toISOString()}
+            pageUpdatedBy={firstUserId}
           />
-        </div>
+        </CardPropertiesWrapper>
       </Paper>
     </Context>
   );
 }
 
-DatabaseKanbanView.parameters = {
+CardPropsView.parameters = {
   msw: {
     handlers: {
       pages: rest.get('/api/spaces/:spaceId/pages', (req, res, ctx) => {
-        return res(ctx.json([boardPage, page1, page2]));
-      }),
-      getBlock: rest.get('/api/blocks/:pageId', (req, res, ctx) => {
-        return res(ctx.json(reduxStore.getState().cards.cards[req.params.pageId as string]));
-      }),
-      updateBlocks: rest.put('/api/blocks', async (req, res, ctx) => {
-        return res(ctx.json(req.json()));
-      }),
-      views: rest.get('/api/blocks/:pageId/views', (req, res, ctx) => {
-        return res(ctx.json([view]));
+        return res(ctx.json([boardPage, page1]));
       })
     }
   }
