+++ stories/Databases/TableProperties.stories.tsx
@@ -4,11 +4,12 @@ import { rest } from 'msw';
 import type { ReactNode } from 'react';
 import { useRef } from 'react';
 import { Provider } from 'react-redux';
+import type { MockStoreEnhanced } from 'redux-mock-store';
 import { v4 as uuid } from 'uuid';
 
-import CardDetailProperties from 'components/common/BoardEditor/focalboard/src/components/cardDetail/cardDetailProperties';
+import Table from 'components/common/BoardEditor/focalboard/src/components/table/table';
+import type { RootState } from 'components/common/BoardEditor/focalboard/src/store';
 import { mockStateStore } from 'components/common/BoardEditor/focalboard/src/testUtils';
-import { CardPropertiesWrapper } from 'components/common/CharmEditor/CardPropertiesWrapper';
 import type { ICurrentSpaceContext } from 'hooks/useCurrentSpace';
 import { CurrentSpaceContext } from 'hooks/useCurrentSpace';
 import { MembersProvider } from 'hooks/useMembers';
@@ -20,14 +21,15 @@ import { createMockBoard, createMockCard } from 'testing/mocks/block';
 import { createMockPage } from 'testing/mocks/page';
 import { generateSchemasForAllSupportedFieldTypes } from 'testing/publicApi/schemas';
 
-import { spaces } from '../../../../.storybook/lib/mockData';
+import { spaces } from '../../.storybook/lib/mockData';
 
 export default {
   title: 'Databases/Composites',
-  component: CardDetailProperties
+  component: Table
 };
 
 const firstUserId = uuid();
+const secondUserId = uuid();
 
 const space = spaces[0];
 
@@ -69,7 +71,7 @@ const card1 = createMockCard(board);
 card1.fields.properties = {
   [schema.text.id]: 'First',
   [schema.checkbox.id]: 'true',
-  [schema.date.id]: '{"from":"1695067400713"}',
+  [schema.date.id]: '{"from":1694001600000}',
   [schema.email.id]: 'test1@example.com',
   [schema.multiSelect.id]: [schema.multiSelect.options[0].id, schema.multiSelect.options[1].id],
   [schema.number.id]: 7223,
@@ -78,10 +80,32 @@ card1.fields.properties = {
   [schema.select.id]: schema.select.options[0].id,
   [schema.url.id]: 'https://www.google.com'
 };
+
+const card2 = createMockCard(board);
+
+card2.fields.properties = {
+  [schema.text.id]: 'Second',
+  [schema.checkbox.id]: 'false',
+  [schema.date.id]: '{"from":1694501600000}',
+  [schema.email.id]: 'test2@example.com',
+  [schema.multiSelect.id]: [schema.multiSelect.options[1].id, schema.multiSelect.options[2].id],
+  [schema.number.id]: 8345,
+  [schema.person.id]: secondUserId,
+  [schema.phone.id]: '+1 (345) 8345 345',
+  [schema.select.id]: schema.select.options[1].id,
+  [schema.url.id]: 'https://www.example.com'
+};
+
 const page1 = createMockPage({
   id: card1.id,
   type: 'card',
   title: 'Card 1'
+});
+
+const page2 = createMockPage({
+  id: card2.id,
+  type: 'card',
+  title: 'Card 2'
 });
 
 const reduxStore = mockStateStore([], {
@@ -100,11 +124,12 @@ const reduxStore = mockStateStore([], {
   cards: {
     current: '',
     cards: {
-      [card1.id]: card1
+      [card1.id]: card1,
+      [card2.id]: card2
     },
     templates: {}
   }
-});
+}) as MockStoreEnhanced<Pick<RootState, 'boards' | 'views' | 'cards'>>;
 
 function Context({ children }: { children: ReactNode }) {
   // mock the current space since it usually relies on the URL
@@ -126,7 +151,11 @@ function Context({ children }: { children: ReactNode }) {
   );
 }
 
-export function CardPropsView() {
+function voidFunction() {
+  return Promise.resolve();
+}
+
+export function DatabaseTableView() {
   // const dispatch = useAppDispatch();
 
   // useEffect(() => {
@@ -136,28 +165,44 @@ export function CardPropsView() {
   return (
     <Context>
       <Paper>
-        <CardPropertiesWrapper>
-          <CardDetailProperties
+        <div className='focalboard-body'>
+          <Table
             board={board}
+            showCard={voidFunction}
             readOnly={false}
+            readOnlySourceData={false}
             views={[view]}
             activeView={view}
-            card={card1}
-            cards={[card1]}
-            pageUpdatedAt={new Date().toISOString()}
-            pageUpdatedBy={firstUserId}
+            addCard={voidFunction}
+            cardIdToFocusOnRender=''
+            onCardClicked={voidFunction}
+            selectedCardIds={[]}
+            visibleGroups={[]}
+            cardPages={[
+              { card: card1, page: page1 },
+              { card: card2, page: page2 }
+            ]}
           />
-        </CardPropertiesWrapper>
+        </div>
       </Paper>
     </Context>
   );
 }
 
-CardPropsView.parameters = {
+DatabaseTableView.parameters = {
   msw: {
     handlers: {
       pages: rest.get('/api/spaces/:spaceId/pages', (req, res, ctx) => {
-        return res(ctx.json([boardPage, page1]));
+        return res(ctx.json([boardPage, page1, page2]));
+      }),
+      getBlock: rest.get('/api/blocks/:pageId', (req, res, ctx) => {
+        return res(ctx.json(reduxStore.getState().cards.cards[req.params.pageId as string]));
+      }),
+      updateBlocks: rest.put('/api/blocks', async (req, res, ctx) => {
+        return res(ctx.json(req.json()));
+      }),
+      views: rest.get('/api/blocks/:pageId/views', (req, res, ctx) => {
+        return res(ctx.json([view]));
       })
     }
   }
